+++ test/Atc.Wpf.Theming.Tests/GlobalUsings.cs
@@ -1,3 +1,11 @@
+global using System.Globalization;
 global using System.Reflection;
+global using System.Windows;
+global using System.Windows.Data;
+global using System.Windows.Media;
 
+global using Atc.Wpf.Theming.Controls.Selectors;
+global using Atc.Wpf.Theming.ValueConverters;
 global using Atc.XUnit;
+
+global using Xunit;